[CITE-177] Addressed PR comments
